+++ src/scope/network/network.js
@@ -5,13 +5,14 @@ import type { ScopeDescriptor } from '../scope';
 import Component from '../../consumer/component';
 import type { ListScopeResult } from '../../consumer/component/components-list';
 import DependencyGraph from '../graph/scope-graph';
+import type { SSHConnectionStrategyName } from './ssh/ssh';
 
 export interface Network {
   connect(host: string): Network;
   close(): void;
   describeScope(): Promise<ScopeDescriptor>;
   fetch(bitIds: BitIds): Promise<ComponentObjects[]>;
-  list(namespacesUsingWildcards?: string): Promise<ListScopeResult[]>;
+  list(namespacesUsingWildcards?: string, strategiesNames?: SSHConnectionStrategyName[]): Promise<ListScopeResult[]>;
   search(query: string, reindex: boolean): Promise<string>;
   show(bitId: BitId): Promise<?Component>;
   deprecateMany(ids: string[], context: ?Object): Promise<Object[]>;