+++ fixtures/angular/html-deps/another-component.ts
@@ -1,7 +1,0 @@
-import { NgModule, Component } from '@angular/core';
-
-@Component({
-  selector: 'another-component',
-  template: '<div>Hello world</div>'
-})
-export class AnotherComponent {}