+++ fixtures/angular/html-deps/main-component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'main-component',
-  template: '<another-component></another-component>'
-})
-export class MainComponent {}