+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- improve interactive behavior when ssh connection failed
+- improve interactive texts
+
 ## [14.2.4-dev.6] - 2019-08-12
 
 - fix init --bare with name