+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [2.1.1-dev.1] - 2019-08-12
+
+- remove angular dependencies. use typescript compiler to parse Angular Decorators
+
 ## [2.1.0] - 2019-07-18
 
 - propagate from component root dir backwards to find tsconfig.json for Angular projects